uncompress archives in subfolders
if the archive downloaded by ghns has more than a file
inside (either single file or a folder at the root),
uncompress it in a folder called the same as the file
name (without path or extension)

this fixes downloading and installing plasma wallpaper
packages (those that provide multiple resolutions).
I don't know if this behavior would in turn break other
applications, but i don't think so as it seems the
behavior changed between kde4 and kf5.
if needed i can make it do this only in the case of
wallpapers (if (standardResourceDirectory ==
QLatin1String("wallpaper"))) but i think this
should be the correct behavior

REVIEW:128740