Deprecate KNS3::DownloadManager
While it might seem to possibly make sense to extend
KNSCore::DownloadManager, the simplistic nature of the class makes
that an unnecessary thing, especially given that it makes better
logical sense to simply move it. We cannot remove this one entirely,
as it is already exported, so create a KNSCore version and mark this one
deprecated instead. Slight code increase, but simpler end result.