Add new archive management option (subdir) to knsrc
KNewStuff would originally decompress installed archives into where ever
they were put, without being put into a subdirectory. In certain
instances, this would cause issues, but the fix for this issue
introduced issues in e.g. KStars, which was suddenly unable to consume
the data.

This patch reverts the default behaviour to the original behaviour, but
adds the option to allow for archives to be decompressed into a
subdirectory, and defaulting to this behaviour when handling wallpapers
(which is what the original patch was done to handle).

Differential Revision: https://phabricator.kde.org/D3470