Fix set up of the ::Engine when created with an absolute config file path
QStrandardPaths::locate returns {} when an absolute path is passed so we
can't use it to extract the file's baseName.
Also simplifies a bit the logic that was relying on a class attribute that
was only used in that function and we turned into a local variable. It was
adding a ':' that doesn't make sense anymore and I dropped as well.

REVIEW: 129660