Use a single QNAM (and a disk cache) for HTTP jobs
Use a single QNetworkAccessManager instance for all our
HTTP jobs, and also add a simple diskcache to that qnam.
Further ensure there is only a single qnam for the entire
application using kns' http jobs, across all threads
(lock when accessing the qnam). Without this, we are
liable to end up creating and destroying a great many
qnam instances, which certainly is something to try and
avoid.

Differential Revision: https://phabricator.kde.org/D5638