[FEATURE] Option to build & install QCH files with the public API dox
Using the new extra-cmake-modules module ECMAddQch (since 5.36.0)
this adds the option to automatically build and install files
in QCH format with the docs about the public API, which then can be
used e.g. in Qt Assistant, Qt Creator or KDevelop.

Additionally the installed cmake config files will be extended
with a target KF5NewStuffCore_QCH resp. KF5NewStuff_QCH containing
information about how to "link" into the generated QCH files,
which then can be used in the cmake build system of other libraries
building on these libraries, by simply listing the target in "LINK_QCHS"
of their ecm_add_qch() usage.
And respective doxygen tag files with all the metadata about the
generated QCH file and used for the "linking" will be created and
installed.

Pass -DBUILD_QCH=ON to cmake to enable this.