Use the right scope for the installpath variable
Summary:
Only mark package as installed if it was successfully extracted

Don't store all installed files in the knsregistry file

At the moment we were storing every file, for big packages (for example
icon themes like Papyrus) this meant ~5 MiB of paths that had to be loaded
when the engine was created.
This patch makes it so instead of adding all the contents of the directory
it will only list the extracted directories so they can recursively be
removed. I don't think it's a supported workflow to be adding stuff in an
extracted directory anyway.

Test Plan:
Tests pass, tried un/installing massive icon themes. Also massif
reports a much lower memory consumption.

Reviewers: #frameworks, whiting, leinir

Reviewed By: leinir

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D6104