Expand KNewStuff documentation
Add mention of KNewStuffCore and KNewStuffQuick to the base
documentation, and expand the API documentation to mention
some of the new functionality, as well as add some documentation
for things which currently lack it. Prior to this, the
documentation had no mention of the way the framework is now
logically split up.

This deficiency was brought to my attention during a chat with a
GSoC student, who was otherwise in the process of disregarding
KNewStuff and headed towards reimplementing large parts of
KNewStuffCore (as the documentation makes no suggestion this is
now a thing that can be used on its own).

Differential Revision: https://phabricator.kde.org/D5902