Require Kirigami 2.1 instead of 1.0 for KNewStuffQuick
This reduces the amount of things which require the by
now quite out of date Kirigami 1.0 further. This also
introduces a find clause to the CMakeLists.txt (a runtime
find, so we should get no issues as it's informational
rather than a hard require). Preparatory work for
extending the KNewStuff Qt Quick components in the
near future.

Differential Revision: https://phabricator.kde.org/D8111