Don't offer qml plugin as a link target
This removes the KNSQuick plugin as a link target for cmake.
It was never intended to be one, and this was just never
picked up as it didn't cause issues until attempting to build
an app using it on the Windows binary factory. The patch also
switches to using the KDE QML install locations, equalising
it with other Frameworks.

Differential Revision: https://phabricator.kde.org/D12082