Add programmaticaly useful error signalling
Previously, KNSCore only signalled errors using string messages,
which were useful for fault finding, but not useful when
attempting to handle errors during general operation. This patch
adds a signal which is fired at the same time as some error
messages, and which provides error condition information which
is useful in a programmatical fashion.

This patch would allow https://bugs.kde.org/show_bug.cgi?id=395937
to be fixed in a non-hackish fashion.

Differential Revision: https://phabricator.kde.org/D16665