Change default sort order in the download dialog to "Most downloads"
Summary:
BUG: 399163

This changes the default sort ordering of the download dialog to "Most downloads" (in 3 months), because the current problem is that the items sorted by rating are sometimes very old and often don't work anymore.
When you change the sort order to "Most downloads" it shows the items that are actually popular (many downloads in 3 months) and they are are more up to date.

Test Plan: Open a GHNS dialog and see that the default sort ordering of the content is by "Most downloads", and that the "Most downloads" radio button is checked.

Reviewers: #frameworks, #vdg, ngraham

Reviewed By: #vdg, ngraham

Subscribers: ngraham, filipf, #vdg, kde-frameworks-devel, #frameworks

Tags: #frameworks, #vdg

Differential Revision: https://phabricator.kde.org/D17641