Fix semantics for ghns_exclude
After much discussion on the topic, it came to light that the
specific grammar used for the default tag in KNSCore was
incorrect, as it was an action verb (rather than an adjective
the way the metadata tags are supposed to be). While this does
mean that previous versions of Frameworks will not have the
filtering functionality by default, it is otherwise
non-invasive and causes no side effects apart from the
server-suggested filtering working.

BUG: 402888
Differential Revision: https://phabricator.kde.org/D18038