Filter out invalid content in lists
Prior to this change, if the ocs server being queried returned a
list of entries containing invalid entries (such as happened a
few days ago for the otherwise empty Akonadi email providers
category), the listing function would faithfully return that list
of entries verbatim, which led consumers of that list to insist
there were more entries to fetch, and then try again, and again,
and again. With this patch, only valid entries get forwarded as
results to the consumers (be they knewstuff's own dialogues, or
others, such as Discover or Peruse).

This change is in KNewStuff rather than Attica, because Attica
by design does not apply heuristics of any sort to its
functions. That, again by design, is what KNewStuff is for.

Differential Revision: https://phabricator.kde.org/D18813