Handle unsupported OCS commands, and don't over-vote
When the server does not support certain commands (such as voting),
don't be quite so loud about it. It is not a critical error, and we
should accept it when we are told so.

Also, the same bug report covers votes being sent out during listing
of items. This is caused by the rating being set sometime after
setting up a signal chain, and then changing the rating of the
widget. So, block signals when we're simply updating the rating's
internal state.

BUG:391111
Differential Revision: https://phabricator.kde.org/D19858