Remove pixelated border
Summary:
In the before-times, drop shadows were considered vitally important on everything, and so it was introduced to KNewStuff's dialogues as well. It was, however, never done quite right, and it's making our dialogues look all silly. So, we remove it for now, and if we still want it, we can reintroduce them in a more modern fashion.
BUG:391108

Test Plan:
Open the details dialogue of any knewstuff listing
Before patch: See pixelated (and weirdly positioned) pixelated border on all preview images
After patch: See no pixelated border

Reviewers: #knewstuff, ngraham, sitter, #vdg

Reviewed By: ngraham, #vdg

Subscribers: anthonyfieroni, mmustac, ndavis, kde-frameworks-devel, #knewstuff

Tags: #frameworks, #knewstuff

Differential Revision: https://phabricator.kde.org/D20693