Ensure that the changedEntries property is correctly propagated
Summary:
Without this change, for some reason unknown to me, the alias
seems to fail quietly (that is, you can still read the property,
but it does not forward the changed signal).

This was discovered during the course of reviewing D24942.

Expose the amount of changed entries

Since changedEntries is only a forwarding type, and not a true list
in the QML sense, there's no good way to get the number of entries
(without for e.g. changing changedEntries to a model, which just
seems like overkill)

Reviewers: #frameworks, #knewstuff, apol, ngraham

Reviewed By: ngraham

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D25041