Use the pointing hand cursor for the single-clickable delegates
Summary:
When you single-click on a delegate, it takes you to the page for that item. Because the delegates look selectable (they have a selection highlight when hovered over), you expect that a single-click will select the item rather than go into it. To make the enter-on-click behavior less unexpected, the cursor should use the finger pointer when hovering over any of the single-clickable delegates.

BUG:413435

Reviewers: #knewstuff, #frameworks, ngraham

Reviewed By: ngraham

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D25168