Show a "Loading more..." busy indicator when loading view data
Summary:
Add an isFetching property to the QML ItemsModel, and use that
to show a busy indicator with a label at the end of the view
of entries when scrolling through it. This ensures the user is
informed when fetching is happening, without being too in-your-
face about it all.

Also remove the passive notifications for busy messages:
Since all the busy messages are for the model, and we now have a
different way to show this to the user, get rid of that one.
We still have a number of idle messages and other such informational
bits, but this gets rid of a great many of the more obnoxious
messages that keep popping up.

Test Plan:
Without patch: A bunch of passive notifications when scrolling through the list, and on first start
With patch: Much fewer of the status-ey messages, and a generally more pleasant feel to the scrolling

Reviewers: #knewstuff, #frameworks, #plasma, ngraham

Reviewed By: ngraham

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D25197