Send along the config name with requests
Summary:
This allows the server to perform some minimal analysis on the
origin of requests, without sending out anything that is in
any real way useful for guessing who is on the requesting end.

The end goal is to work out which bits of our software is causing
the largest amount of pressure on the servers, so we can work out
where to expend effort with the highest amount of impact.

Reviewers: #knewstuff, #frameworks, #plasma, starbuck, ronaldv, mart

Reviewed By: #plasma, mart

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D25441