Fix the KNSQuick Dialog's close title and basic layout
Summary:
This resolves several issues found in the KNSQuick Dialog component.

Specifically it puts a background in the dialog (as a rectangle),
since the QtQuick Dialogs base dialog lacks such a thing by default,
and offers no background property to rectify it.

It further changes the Close button from a custom one to using a
default button provided by the various underlying systems. This
is more fault proof than the previous incarnation, but requires
that the escape button press is handled separately.

Finally it sets a sensible default size (previously the dialog
was made large by the layout, but as a side effect of the layouting
now being done without a Layout (which we could, but which would make
the code enormously convoluted for little gain), we can now set a
set of implicit width and height on the content item, which results
in both a sensible default size and a dialogue that will resize down
the way one would expect it to).

This could have been done as separate patches, but they're too
intertwined for that to feel in any way reasonable to do...

BUG:415606
BUG:416354
BUG:414682

Test Plan:
The dialog with its new default size looks as follows:
{F8089818}

Reviewers: #frameworks, ngraham

Reviewed By: ngraham

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D27204

(cherry picked from commit fd5b37818e6aa72fc149d9aec58511ca83e8a9d4)