Fix clicking thumb-only delegate
Summary:
This fixes a call to an attached property which only exists on the
delegate root, not the children.

BUG:418368

Test Plan: Switch to the thumbnail only view and click an item. Notice that without the patch the click is ignored (and produces error output on the console), and that with the patch it works.

Reviewers: #frameworks, ngraham

Reviewed By: ngraham

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D27777