[QtQuick dialog] Port to UrlBUtton and hide when there's no URL
Summary:
Port from `linkButton` to `UrlButton` for links that open webpages, as this is what
`UrlButton` was specifically designed for. Also hide any link that doesn't actually go
anywhere because the URL isn't defined by the data source.

Test Plan:
{F8211891}

There's no longer a donation button that doesn't do anything when clicked, and the
homepage link shows a tooltip when hovered.

Reviewers: leinir, #knewstuff, #vdg, apol

Reviewed By: apol

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D28522