Introduce more user-visible error reporting for installations
Summary:
Prior to this, we did have error reporting, but only in the way of
writing errors out on the command line (through qCCritical). While
this is fine for debugging purposes, it is unfortunate when errors
can be expected to occur in day to day use (as we are dealing with
a wide range of things that can (and often do) go wrong), and giving
proper feedback to the user as to why, for example, their new icons
aren't showing up when they seem to have installed just fine would
seem a reasonable thing to do.

* Add a signal to Installation fired when there's a critical error
* Forward the installer errors through KNSCore::Engine
* Add the MessageBoxSheet component from kaccounts-integration
* Add a component for displaying errors from the engine
* Use the ErrorDisplayer component (only show on the current page)

CCBUG:418466

Test Plan:
An error displayey thing:
{F8220509}

Reviewers: #knewstuff, #plasma, ngraham, #frameworks

Reviewed By: ngraham

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D28532