Fix update auto selection
Summary:
The new autoselection for updating installed entries would firstly
not be looking at the right amount of download links, and further it
would not be able to identify the links it needed to, as it did not
look at the descriptive names (where the filename is often stored),
which then would not match. These changes ensure that we look at
the right bits of information when doing this check. End result is
that updating works.

* Ensure that the cache doesn't forget installing and updating items
* Analyse more of the download link information, and use correct counts

BUG:419959

Test Plan:
Find an item in some KNS dialog in need of updating, and click the Update button

On the console, when launched using the correct logging categories, will
tell you about the selection process (and the UI will show that an update
is in progress, and eventually that it is completed). To do this check with
the icons kcm, run it like so:

  QT_LOGGING_RULES="org.kde.knewstuff*=true" kcmshell5 kcm_icons

Reviewers: #frameworks, #plasma, bugseforuns, ngraham, mart

Reviewed By: #plasma, mart

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D29222