KNS: Deprecate isRemote method and handle parse error properly
Summary:
The isRemote check and the FIXMEs are removed, instead the result from the empty target check is properly handled.

The same check as in isRemote is done in the Installation::readConfig method (line 122).
If this fails a warning is given and in the engine is an error emitted. This error gets shown in the GUI.
Then there are no other checks needed.

Test Plan:
Delete the line TargetDir from the dolphin servicemenu knsrc file.
(Dolphin repository src/settings/services/servicemenu.knsrc).

Before:
{F8256199}
After, an error message is shown:
{F8256197}

Reviewers: #knewstuff, ngraham, leinir

Reviewed By: #knewstuff, leinir

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D29455