Fix showing updates when the option is selected
Summary:
This patch primarily fixes the issue that when picking the "Updateable Only" option on NewStuff.Page, we would be shown all the entries anyway. It also fixes duplicated entries on the "Installed Only" option, and further adds a little placeholder for when the view is expectedly empty.

* Use the right connections for update entry display
* While potentially expensive, only add entries to the model once
* Use Kirigami.PlaceholderMessage to show we have no entries when we don't

BUG:416762

Test Plan:
Without this patch, behaviour is as shown in https://bugs.kde.org/show_bug.cgi?id=416762

With this patch, the behaviour is as expected (shows only updateable entries with that option selected, only shows installed entries once with that option, and shows a useful message when the list is empty and expected to be empty)

Reviewers: #knewstuff, #plasma, bugseforuns, ngraham, #frameworks

Reviewed By: ngraham

Subscribers: alex, kde-frameworks-devel

Tags: #frameworks, #plasma, #knewstuff

Differential Revision: https://phabricator.kde.org/D29447