KNS: Do not mark entry as installed if install script failed
Summary:
If the install script failed/was aborted the entry gets marked as installed.
Now the exit code is checked and a error message is displayed.

This is the counterpart of {D29123}.

Test Plan:
Make sure to have {D29119}. Install the deb package of the "Jetbrains" dolphin addon.
When the authorization popup comes press escape. The dolphin installer
shows an error popup and the entry is not marked as installed.
If you are not on a debian based system it should throw an error anyway ;-).

Reviewers: #knewstuff, ngraham, leinir

Reviewed By: #knewstuff, leinir

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D29451