Fix incorrect colours in the KNS Quick messagebox
Summary:
Explicitly set the colours used in the messagebox' content item,
as the theme does not propagate to the simpler items, only to
those based on top of QtQuick Controls.

BUG:421270

Test Plan: Ensure you are using a dark colour scheme. Provoke an error dialog in any KNewStuff Quick dialog, and see that this patch makes the error message colour turn from a near-invisible text on a dark background (as shown in https://bugs.kde.org/show_bug.cgi?id=421270) to a visible text on a dark background.

Reviewers: #frameworks, #knewstuff, #plasma, bugseforuns, ngraham

Reviewed By: ngraham

Subscribers: ngraham, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D29678