Don't allow content to overflow in the grid delegates
Due to the content shown in the delegates being user created, we
cannot trust that it will flow correctly when shown inside a
Text (or Label) Component. Consequently, we have to deal with the
unfortunate cases where that content would end up overflowing.
While it is suboptimal to do this by just clipping it, it will
make the delegates look less broken, and we can then suggest
to the content creator to maybe consider the presentation
of their content in KNS as well as on the website.

BUG:422476