Remove some old testing code
This line used to set the sorting mode to an unexpected "most
downloaded", when the rest of KNewStuff expects it to be "most recent".
This did expose the bug this patch otherwise fixes, but also meant that
the code was out of sync with the rest of KNewStuff, which we won't
want.