Fix details install dropdown in QWidgets dialog
BUG: 369561
FIXED-IN: 5.73

The button was connected to the install slot, but if there
is a dropdown the actions get the signal emitted and this signal
was not handled.

To try this out you can search for "jetbrains" in the dolphin
services KNS dialog and go to the details page and install one of
the variants.