Add a RemoveDeadEntries option for knsrc files
If you set RemoveDeadEntries=true, entries whose installed files
have all been deleted without going through KNewStuff will be removed
from the cache. The removal will happen if, and only if, all listed
files were removed, which means that if, for example, an entry was
installed from archive, which was decompressed to yield multiple
installed files, if even one of those files remains, the entry will
remain marked as installed.

This is to help work around people deleting files from underneath
KNewStuff, which turns out to happen a lot with e.g. wallpapers and icons.

CCBUG:417985

Only CCBUGing this one, as it is an imcomplete fix for the issue,
and requires intervention in the icons.knsrc file for the fix to
be applied to the issue mentioned in that bug.