Remove * character when passing dir to script
For the script to be called the file must exist
and if the path ends with * this always returns not true.
We have to remove this last character (the meaning of *
is that all the subdirs/subfiles are part of the installed files)
and check if the dir then exists.

In that case we can also get rid of the manuallyUninstalled check
which has the same problem, @leinir has implemented
another solution for this which we can apply.