Hide load more spinner while updating/installing
BUG: 422047
FIXED-IN: 5.73

The isLoadingData state was always set to true when the busy signal gets
emitted. But in the case of an installation there is no data in the
sense of more entries loaded, consequently we don't need to show the
spinner.

Because of that this MR introduces a new API which tells exactly what operations are currently running.