Run uninstall script async
CCBUG: 418042

Otherwise the UI would be blocked. In some cases this is not notable,
because we have a dialog open. But especially if we uninstall a package
using packagekit the UI is unresponsive for a little while.

Maybe we should put the deletion of the files in a seperate job
or looking towards KF6 at least in a private impl but for now a
lambda is the best we can do.