Disable search when not available
When the user wants to view the installed or updatable entries
the search should get disabled. This behavior is the same in
the old QWidgets dialog.

Otherwise if we were going to type a query while in the installed/updatable
only mode we would get the exact same results after submitting the query, which is confusing.