Add a reverse lookup for entries based on their installed files
This is a step towards allowing for much deeper integration of KNewStuff
into various parts of the frontends which make use of it. As an example,
we want to display updateability status for entries in Plasma's widget
explorer, and we want to show next/previous series details for Peruse's
downloaded books. This will allow us to do so.