Make the internal kpackage job task less fragile
This is done to ensure we don't end up emitting things on an
already-deleted object, which breaks. We cannot simply swap the existing
code out for a connection that's queued, as we may end up not emitting
the result, which would be a problem in itself.

Also clean up the error reporting a touch while we're at it.

BUG:425811