Fix cache sync issues with QtQuick dialog
The engine is reused for the QtQuick dialog, consequently
the cache gets only updated once the dialog is initialized.
Now the cache is updated if the RemoveDeadEntries option is selected
every time every time the dialog is about to show.

Also the state of the entries gets updated and with
8531bbcc
the current filter rev-alidates these entries properly.

BUG: 417985
FIXED-IN: 5.74